minor fix for hash event
