--- instrumentation/csec_crypto/go.md
@@ -1,1 +1,0 @@
-module github.com/newrelic/csec-go-agent/instrumentation/csec_crypto
