+++ instrumentation/csec_crypto/go.mod
@@ -1,0 +1,3 @@
+module github.com/newrelic/csec-go-agent/instrumentation/csec_crypto
+
+go 1.23.3
