+++ instrumentation/csec_crypto/sec_blake2b.go
@@ -24,7 +24,7 @@ func secBlake2sSum256_s(data []byte) [blake2s.Size]byte {
 	return secBlake2sSum256_s(data)
 }
 
-func PluginStart_blake2b() {
+func initblake2sHook() {
 
 	e := secIntercept.HookWrap(blake2s.Sum256, secBlake2sSum256, secBlake2sSum256_s)
 	secIntercept.IsHookedLog("blake2s.Sum256", e)
